requirements.txt
bleach~=6.0.0

mysql-connector-python >= 8.0.28, <=8.0.29

clamd==1.0.2

django-mysql>=3.10.0,<3.11.0

Django>=3.2,<4.0

djangorestframework>=3.12.2,<3.13.0

flake8>=3.9.0,<4.0.1

gunicorn>=20.0.4,<20.1.0

coverage>=5.5,<6.0

django-cors-headers>=3.7.0,<3.8.0

ddt>=1.4.2,<1.5.0

django-model-utils>=4.1.1,<4.2.0

isort>=5.11.0, <5.12.0

openlxp-authentication >=1.1.0, <1.2

python-magic >=0.4.27, <0.5

requests>=2.25.1,<2.26.0

django-admin-interface>=0.28.3

django-colorfield>=0.11.0

python-slugify>=8.0.1

text-unidecode>=1.3

neomodel>=5.1.0,<6.0.0

pandas>=0.23.4

django-neomodel>=0.1.1

sentence-transformers>=2.0.0
